Moving test to tests/it
