+++ src/array/utf8/json.rs
@@ -11,7 +11,6 @@ use crate::{
 };
 
 /// Implements json deserialization from a Utf8Array
-#[cfg(feature = "io_json")]
 impl<O: Offset> Utf8Array<O> {
     /// Infers the DataType from a number of JSON rows in a Utf8Array
     pub fn json_infer(&self, number_of_rows: Option<usize>) -> Result<DataType> {
@@ -52,33 +51,5 @@ impl<O: Offset> Utf8Array<O> {
             .collect::<Result<Vec<Value>>>()?;
 
         Ok(_deserialize(&rows, data_type))
-    }
-}
-
-#[cfg(all(test, feature = "io_json"))]
-mod tests {
-
-    use super::*;
-    use crate::array::{Array, StructArray};
-
-    #[test]
-    fn json_deserialize() {
-        let array = Utf8Array::<i64>::from([
-            Some(r#"{"a": 1, "b": [{"c": 0}, {"c": 1}]}"#),
-            None,
-            Some(r#"{"a": 2, "b": [{"c": 2}, {"c": 5}]}"#),
-            None,
-        ]);
-        let data_type = array.json_infer(None).unwrap();
-        let new_array = array.json_deserialize(data_type).unwrap();
-
-        // Explicitly cast as StructArray
-        let new_array = new_array.as_any().downcast_ref::<StructArray>().unwrap();
-
-        assert_eq!(array.len(), new_array.len());
-        assert_eq!(array.null_count(), new_array.null_count());
-
-        let field_names: Vec<String> = new_array.fields().iter().map(|f| f.name.clone()).collect();
-        assert_eq!(field_names, vec!["a".to_string(), "b".to_string()]);
     }
 }
